refactor constants, mdx imports
