+++ src/docs/constants/docsItems.ts
@@ -1,3 +1,5 @@
+import { componentTypes } from "@examples";
+
 export type DocItems = DocItemGroup[];
 export type DocItemGroup = {
   label: string;
@@ -7,42 +9,6 @@ export type DocItem = {
   href: string;
   label: string;
 };
-
-// Temp fix duplication, when imported from src/docs/examples/index.ts it cause import errors through dynamic import of examples
-const componentTypes = [
-  {
-    type: "hero-sections",
-    title: "Hero section",
-  },
-  {
-    type: "lists",
-    title: "Lists",
-  },
-  {
-    type: "badges",
-    title: "Badges",
-  },
-  {
-    type: "buttons",
-    title: "Buttons",
-  },
-  {
-    type: "emojis",
-    title: "Emojis",
-  },
-  {
-    type: "social-icons",
-    title: "Social icons",
-  },
-  {
-    type: "texts",
-    title: "Paragraphs",
-  },
-  {
-    type: "dividers",
-    title: "Dividers",
-  },
-];
 
 export const docsItems: DocItems = [
   {
